# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Discord Bot Hosting Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files and folders
      security:
        - bearerAuth: []
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: "/"
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileList"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or folder
      security:
        - bearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /files/content:
    get:
      operationId: getFileContent
      tags: [files]
      summary: Get file content for editing
      security:
        - bearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
    put:
      operationId: updateFileContent
      tags: [files]
      summary: Save file content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileContentUpdate"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /files/create:
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file or folder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
  /files/rename:
    post:
      operationId: renameFile
      tags: [files]
      summary: Rename a file or folder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRename"
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /files/move:
    post:
      operationId: moveFile
      tags: [files]
      summary: Move a file or folder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileMove"
      responses:
        "200":
          description: Moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /files/copy:
    post:
      operationId: copyFile
      tags: [files]
      summary: Copy a file or folder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileCopy"
      responses:
        "201":
          description: Copied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /logs:
    get:
      operationId: getLogs
      tags: [logs]
      summary: Get bot logs
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
        - name: level
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
    delete:
      operationId: clearLogs
      tags: [logs]
      summary: Clear all logs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    FileItem:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, folder]
        size:
          type: ["integer", "null"]
        modified_at:
          type: ["string", "null"]
          format: date-time
        extension:
          type: ["string", "null"]
      required: [name, path, type]
    FileList:
      type: object
      properties:
        path:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
        total_files:
          type: integer
        total_folders:
          type: integer
      required: [path, items, total_files, total_folders]
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        encoding:
          type: string
        size:
          type: integer
        modified_at:
          type: ["string", "null"]
          format: date-time
      required: [path, content, encoding, size]
    FileContentUpdate:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required: [path, content]
    FileCreate:
      type: object
      properties:
        path:
          type: string
        type:
          type: string
          enum: [file, folder]
        content:
          type: string
      required: [path, type]
    FileRename:
      type: object
      properties:
        path:
          type: string
        new_name:
          type: string
      required: [path, new_name]
    FileMove:
      type: object
      properties:
        source:
          type: string
        destination:
          type: string
      required: [source, destination]
    FileCopy:
      type: object
      properties:
        source:
          type: string
        destination:
          type: string
      required: [source, destination]
...
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        level:
          type: string
        message:
          type: string
        source:
          type: string
      required: [id, timestamp, level, message, source]
...